+++ src/Sulu/Bundle/WebsiteBundle/Sitemap/SitemapProviderPool.php
@@ -73,28 +73,6 @@ class SitemapProviderPool implements SitemapProviderPoolInterface
     /**
      * {@inheritdoc}
      */
-    public function getFirstAlias()
-    {
-        return reset($this->aliases);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function needsIndex()
-    {
-        return 1 < count($this->providers)
-            || 1 < array_reduce(
-                $this->getIndex(),
-                function ($v1, Sitemap $v2) {
-                    return $v1 + $v2->getMaxPage();
-                }
-            );
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function getIndex()
     {
         if ($this->index) {