+++ CHANGELOG.md
@@ -17,6 +17,7 @@ CHANGELOG for Sulu
     * FEATURE     #2979 [MediaBundle]         Hide internal formats in UI
     * BUGFIX      #2978 [ContentBundle]       Fixed proxy-factory configuration for smart-content
     * FEATURE     #2972 [MediaBundle]         Focus point for media cropping
+    * BUGFIX      #2981 [WebsiteBundle]       Fixed dumping of sitemaps
     * BUGFIX      #2970 [MediaBundle]         Resize image in media edit overlay
     * FEATURE     #2967 [ContentBundle]       Refactored link-tag to allow extending over provider
     * FEATURE     #2966 [MediaBundle]         Added warning when unsaved crop will be lost